fix: graph caching and throttling
