+++ packages/cysync-core/src/components/Topbar.tsx
@@ -51,7 +51,7 @@ export interface TopbarProps {
   tag?: TopbarUIProps['tag'];
 }
 
-export const Topbar: FC<TopbarProps> = props => {
+const TopbarComponent: FC<TopbarProps> = props => {
   const dispatch = useAppDispatch();
   const { lang, discreetMode, accountSync } = useAppSelector(selector);
   const { connection } = useDevice();
@@ -92,8 +92,10 @@ export const Topbar: FC<TopbarProps> = props => {
   );
 };
 
-Topbar.defaultProps = {
+TopbarComponent.defaultProps = {
   icon: undefined,
   subTitle: undefined,
   tag: undefined,
 };
+
+export const Topbar = React.memo(TopbarComponent);
