+++ packages/cysync-core/src/hooks/useGraph/types.ts
@@ -1,0 +1,5 @@
+export interface UseGraphProps {
+  assetId?: string;
+  parentAssetId?: string;
+  accountId?: string;
+}
