+++ packages/cysync-core/src/utils/memoize.ts
@@ -1,0 +1,22 @@
+import lodash from 'lodash';
+
+import { getObjectId } from './objectId';
+
+export function memoizeFunctionWithObjectArg<T extends (arg: any) => any>(
+  func: T,
+) {
+  return lodash.memoize(func, params => {
+    let id = '';
+    for (const val of Object.values(params)) {
+      if (!val) {
+        id += 'null-';
+      } else if (typeof val === 'object') {
+        id += `${getObjectId(val as any)}-`;
+      } else {
+        id += `${val}-`;
+      }
+    }
+
+    return id;
+  });
+}
